# Make sum along dims of SymTridiagonal, Tridiagonal, Bidiagonal, Diagonal faster