--- a/stdlib/LinearAlgebra/src/bidiag.jl
+++ b/stdlib/LinearAlgebra/src/bidiag.jl
@@ -876,3 +876,33 @@ end
 eigen(M::Bidiagonal) = Eigen(eigvals(M), eigvecs(M))
 
 Base._sum(A::Bidiagonal, ::Colon) = sum(A.dv) + sum(A.ev)
+function Base._sum(A::Bidiagonal, dims::Integer)
+    res = Base.reducedim_initarray(A, dims, zero(eltype(A)))
+    @inbounds begin
+        if (dims == 1 && A.uplo == 'U') || (dims == 2 && A.uplo == 'L')
+            res[1] = A.dv[1]
+            for i = 2:length(A.dv)
+                res[i] = A.ev[i-1] + A.dv[i]
+            end
+        elseif (dims == 1 && A.uplo == 'L') || (dims == 2 && A.uplo == 'U')
+            for i = 1:length(A.dv)-1
+                res[i] = A.ev[i] + A.dv[i]
+            end
+            res[end] = A.dv[end]
+        elseif dims >= 3
+            if A.uplo == 'U'
+                for i = 1:length(A.dv)-1
+                    res[i,i]   = A.dv[i]
+                    res[i,i+1] = A.ev[i]
+                end
+            else
+                for i = 1:length(A.dv)-1
+                    res[i,i]   = A.dv[i]
+                    res[i+1,i] = A.ev[i]
+                end
+            end
+            res[end,end] = A.dv[end]
+        end
+    end
+    res
+end