--- a/stdlib/LinearAlgebra/src/bidiag.jl
+++ b/stdlib/LinearAlgebra/src/bidiag.jl
@@ -149,6 +149,9 @@ end
 function Matrix{T}(A::Bidiagonal) where T
     n = size(A, 1)
     B = zeros(T, n, n)
+    if n == 0
+        return B
+    end
     for i = 1:n - 1
         B[i,i] = A.dv[i]
         if A.uplo == 'U'