--- a/stdlib/LinearAlgebra/test/diagonal.jl
+++ b/stdlib/LinearAlgebra/test/diagonal.jl
@@ -647,7 +647,7 @@ end
     @test E.vectors == [0 1 0; 1 0 0; 0 0 1]
 end
 
-@testset "sum" begin
+@testset "sum, mapreduce" begin
     D = Diagonal([1,2,3])
     Ddense = Matrix(D)
     @test sum(D) == 6
@@ -655,6 +655,12 @@ end
     @test sum(D, dims=1) == sum(Ddense, dims=1)
     @test sum(D, dims=2) == sum(Ddense, dims=2)
     @test sum(D, dims=3) == sum(Ddense, dims=3)
+    @test mapreduce(one, min, D, dims=1) == mapreduce(one, min, Ddense, dims=1)
+    @test mapreduce(one, min, D, dims=2) == mapreduce(one, min, Ddense, dims=2)
+    @test mapreduce(one, min, D, dims=3) == mapreduce(one, min, Ddense, dims=3)
+    @test mapreduce(zero, max, D, dims=1) == mapreduce(zero, max, Ddense, dims=1)
+    @test mapreduce(zero, max, D, dims=2) == mapreduce(zero, max, Ddense, dims=2)
+    @test mapreduce(zero, max, D, dims=3) == mapreduce(zero, max, Ddense, dims=3)
 end
 
 @testset "logabsdet for generic eltype" begin