--- a/stdlib/LinearAlgebra/test/tridiag.jl
+++ b/stdlib/LinearAlgebra/test/tridiag.jl
@@ -477,8 +477,20 @@ end
 end
 
 @testset "sum" begin
-    @test sum(Tridiagonal([1,2], [1,2,3], [7,8])) == 24
-    @test sum(SymTridiagonal([1,2,3], [1,2])) == 12
+    T = Tridiagonal([1,2], [1,2,3], [7,8])
+    Tdense = Matrix(T)
+    S = SymTridiagonal([1,2,3], [1,2])
+    Sdense = Matrix(S)
+    @test sum(T) == 24
+    @test sum(S) == 12
+    @test_throws ArgumentError sum(T, dims=0)
+    @test sum(T, dims=1) == sum(Tdense, dims=1)
+    @test sum(T, dims=2) == sum(Tdense, dims=2)
+    @test sum(T, dims=3) == sum(Tdense, dims=3)
+    @test_throws ArgumentError sum(S, dims=0)
+    @test sum(S, dims=1) == sum(Sdense, dims=1)
+    @test sum(S, dims=2) == sum(Sdense, dims=2)
+    @test sum(S, dims=3) == sum(Sdense, dims=3)
 end
 
 end # module TestTridiagonal