As discussed in #34553, the return type of a `sum` along a dimension of a `SymTridiagonal`/`Tridiagonal`/`Bidiagonal`/`Diagonal` matrix should be a dense matrix and not a sparse matrix. Additionally, I've added some specialized methods for the `sum` along a dimension of such a structured matrix.

Now, I get the following timings on my notebook for the test problem reported in #34553.
```julia
julia> using LinearAlgebra, BenchmarkTools

julia> N = 1000;

julia> T = Tridiagonal(ones(N-1), ones(N), ones(N-1));

julia> @benchmark sum($T, dims=$1)
BenchmarkTools.Trial: 
  memory estimate:  7.94 KiB
  allocs estimate:  1
  --------------
  minimum time:     638.978 ns (0.00% GC)
  median time:      1.213 μs (0.00% GC)
  mean time:        1.404 μs (6.90% GC)
  maximum time:     18.222 μs (78.77% GC)
  --------------
  samples:          10000
  evals/sample:     92

julia> @benchmark sum($T, dims=$2)
BenchmarkTools.Trial: 
  memory estimate:  8.13 KiB
  allocs estimate:  7
  --------------
  minimum time:     952.500 ns (0.00% GC)
  median time:      1.501 μs (0.00% GC)
  mean time:        1.649 μs (6.04% GC)
  maximum time:     80.741 μs (92.46% GC)
  --------------
  samples:          10000
  evals/sample:     10

julia> @benchmark sum($T', dims=$1)
BenchmarkTools.Trial: 
  memory estimate:  7.98 KiB
  allocs estimate:  2
  --------------
  minimum time:     694.281 ns (0.00% GC)
  median time:      1.178 μs (0.00% GC)
  mean time:        1.308 μs (6.22% GC)
  maximum time:     29.076 μs (84.47% GC)
  --------------
  samples:          10000
  evals/sample:     57
```
These timings are three to four orders of magnitude lower than the ones reported in #34553.

The difference between `dims=1` and `dims=2` is caused by #35199:
```julia
ulia> using LinearAlgebra, BenchmarkTools

julia> N = 1000;

julia> T = Tridiagonal(ones(N-1), ones(N), ones(N-1));

julia> @benchmark Base.reduced_indices($(axes(T)), $1)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.018 ns (0.00% GC)
  median time:      0.021 ns (0.00% GC)
  mean time:        0.023 ns (0.00% GC)
  maximum time:     16.655 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000

julia> @benchmark Base.reduced_indices($(axes(T)), $2)
BenchmarkTools.Trial: 
  memory estimate:  192 bytes
  allocs estimate:  6
  --------------
  minimum time:     263.777 ns (0.00% GC)
  median time:      271.880 ns (0.00% GC)
  mean time:        291.368 ns (2.19% GC)
  maximum time:     6.761 μs (94.72% GC)
  --------------
  samples:          10000
  evals/sample:     346
```