@ranocha 2020-03-23
Yes, `init` is an initialization value in general. Concerning the assignments to `res`, these could be `+=` assignments, since the initialization value is `zero(eltype(A))`, but that wouldn't change the final result (unless `x + zero(x) != x`, which seems to break the contract of `zero`).
I wrote `res[i] = ` instead of `res[i] +=` to make the purpose of `sum` more clear. Shall I change all assignments?