@dkarrasch 2020-03-23
That's why I was asking about the purpose of `init`. Maybe I was confused by ` reduce(op, itr; [init])`, where you can set arbitrary `init`ial values. `sum` appears to have no such option, so if `init` is some sort of `zero` of the appropriate type, then the current PR is good as is.