@ranocha 2020-03-23
Yes, that's right. For `sum`, there is no possibility to choose `init` different from `zero` of the appropriate type.