@tkf 2020-03-23
> `reduce(op, itr; [init])`, where you can set arbitrary `init`ial values

Nit-picking (sorry), but this is not true. `init` must be the identity of `op`. The reason why "it works" is that `reduce` dispatches to `foldl` (where `init` can be anything).

---

But I don't think it matters here.  You are using `Base.reducedim_initarray` you just defined in `Base._sum` with explicit `init` (`zero(eltype(A))`). So, I think it is fine as long as the definition of `Base._sum` is correct.

If you want to be careful maybe you can invoke `mapreduce` with `dims` directly in the test to check other paths that calls `reducedim_initarray`.