@dkarrasch 2020-03-23
True. I overlooked both the comment in the docstring of `reduce` and the call of `Base.reducedim_initarray` with `zero(eltype(A))`. Sorry! I'll better leave review of this PR to others. 🤣