@ranocha 2020-03-23
> If you want to be careful maybe you can invoke `mapreduce` with `dims` directly in the test to check other paths that calls `reducedim_initarray`.

Thanks for this suggestion, @tkf. After doing this, I realized that I need to change another method of `reducedim_initarray` to make it work also for `max` and `min` instead of `+`.